Improved password security
Time to let go of unsalted MD5 hashes ... 🤦
Old user accounts can still be used but you *really* should change the
password once for every user to update to the new algorithm.

Note that the now used password_hash is PHP >= 5.5.
If that's an issue for you, have a look at password_compat.